+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * [karlwilbur] - Reset CAPTCHA code on subsequent requests for a new image/audio file
 * [karlwilbur] - Add motion blur, using same blur values, when blur is used
 * [karlwilbur] - Add support for a range of CAPTCHA characters
+* [karlwilbur] - Updates to `README.md`
 
 ## 0.9.2 (2020-12-16)