+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.10.0 (2021-02-22)
+
+* [karlwilbur] - Use `kandr-pry-plus` gem instead of the stale and broken `pry-plus`
+* [karlwilbur] - Reset CAPTCHA code on subsequent requests for a new image/audio file
+* [karlwilbur] - Add motion blur, using same blur values, when blur is used
+* [karlwilbur] - Add support for a range of CAPTCHA characters
+* [karlwilbur] - Updates to `README.md`
+
 ## 0.9.2 (2020-12-16)
 
 * [karlwilbur] - [gemspec] Fix files included in gem